+++ packages/api/datewise/lib/date-time-range/date-time-range.dto.ts
@@ -4,15 +4,11 @@ import { DateTimeRange } from './date-time-range.js'
 
 export class DateTimeRangeDto extends FilterQuery {
   @TimestampApiProperty({ description: 'start of the range, inclusive' })
-  @IsTimestamp({
-    isBefore: (dto: DateTimeRangeDto) => dto.until
-  })
+  @IsTimestamp({ isBefore: (dto: DateTimeRangeDto) => dto.until })
   from: string
 
   @TimestampApiProperty({ description: 'end of the range, exclusive' })
-  @IsTimestamp({
-    isAfter: (dto: DateTimeRangeDto) => dto.from
-  })
+  @IsTimestamp({ isAfter: (dto: DateTimeRangeDto) => dto.from })
   until: string
 
   parse (): DateTimeRange {