+++ packages/api/nestjs-file-storage/src/providers/s3/s3.ts
@@ -1,14 +1,13 @@
 import { Readable } from 'stream'
-import type { ReadableStream as NodeReadableStream } from 'node:stream/web'
 import { Injectable } from '@nestjs/common'
-import { DeleteObjectCommand, GetObjectCommand, PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
+import { DeleteObjectCommand, GetObjectCommand, ListObjectsV2Command, PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
 import { Upload } from '@aws-sdk/lib-storage'
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
 import { validateSync } from 'class-validator'
 import { plainToClass } from 'class-transformer'
 import { S3Config } from '#src/providers/s3/s3.config.js'
 import { S3_DOWNLOAD_URL_EXPIRES_S, S3_UPLOAD_EXPIRES_MS, S3_UPLOAD_URL_EXPIRES_S } from '#src/providers/s3/s3.constants.js'
-import { FileStorage } from '#src/providers/file-storage-provider.js'
+import { FileIndex, FileStorage } from '#src/providers/file-storage-provider.js'
 
 @Injectable()
 export class S3 extends FileStorage {
@@ -170,9 +169,7 @@ export class S3 extends FileStorage {
     await parallelUploads.done()
   }
 
-  public async downloadStream (
-    key: string
-  ): Promise<Readable> {
+  public async downloadStream (key: string): Promise<Readable> {
     this.validateKey(key)
 
     const command = new GetObjectCommand({
@@ -185,7 +182,7 @@ export class S3 extends FileStorage {
       throw new Error(`Could not download ${key}`)
     }
 
-    return Readable.fromWeb(result.Body.transformToWebStream() as NodeReadableStream)
+    return result.Body as Readable
   }
 
   public async delete (key: string): Promise<void> {
@@ -197,5 +194,23 @@ export class S3 extends FileStorage {
     })
 
     await this.client.send(command)
+  }
+
+  public async list (options?: {
+    prefix?: string
+    startAfter?: string
+  }): Promise<FileIndex[]> {
+    const command = new ListObjectsV2Command({
+      Bucket: this.bucketName,
+      Prefix: options?.prefix,
+      StartAfter: options?.startAfter
+    })
+
+    const result = await this.client.send(command)
+
+    return result.Contents?.map(item => ({
+      key: item.Key,
+      lastModified: item.LastModified
+    })) ?? []
   }
 }