+++ packages/api/nestjs-file-storage/src/providers/test/test-file-storage.ts
@@ -1,6 +1,6 @@
-import { Readable } from 'node:stream'
+import { Readable } from 'stream'
 import { Injectable } from '@nestjs/common'
-import { FileStorage } from '#src/providers/file-storage-provider.js'
+import { FileIndex, FileStorage } from '#src/providers/file-storage-provider.js'
 
 @Injectable()
 export class TestFileStorage extends FileStorage {
@@ -39,7 +39,19 @@ export class TestFileStorage extends FileStorage {
     return Promise.resolve()
   }
 
+  downloadStream (_key: string): Promise<Readable> {
+    const stream = new Readable()
+    stream.push('test')
+    stream.push(null)
+
+    return Promise.resolve(stream)
+  }
+
   delete (_key: string): Promise<void> {
     return Promise.resolve()
+  }
+
+  list (): Promise<FileIndex[]> {
+    return Promise.resolve([])
   }
 }