+++ packages/api/node-doc-processor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @wisemen/node-doc-processor
 
+## 1.2.8
+
+### Patch Changes
+
+- [#793](https://github.com/wisemen-digital/wisemen-core/pull/793) [`46de4ba`](https://github.com/wisemen-digital/wisemen-core/commit/46de4bac68e067a502893133db357b6fc9898381) Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - Rework internal dependencies to non-peer dependencies
+
+- Updated dependencies [[`0f6efb2`](https://github.com/wisemen-digital/wisemen-core/commit/0f6efb266299e7067889a46ef74da93c503a1018), [`46de4ba`](https://github.com/wisemen-digital/wisemen-core/commit/46de4bac68e067a502893133db357b6fc9898381)]:
+  - @wisemen/pgboss-nestjs-job@3.3.3
+
 ## 1.2.7
 
 ### Patch Changes